+++ src/diagnostic/cnt/Container.h
@@ -25,29 +25,29 @@ public:
 		return mUpperBound[i];
 	}
 
-	inline unsigned int depth() const
+	inline size_t depth() const
 	{
 		return mDepth;
 	}
 
-	inline unsigned int nodeCount() const
+	inline size_t nodeCount() const
 	{
 		return mNodeCount;
 	}
 
-    inline unsigned int innerNodeCount() const
+    inline size_t innerNodeCount() const
     {
         return mInnerCount;
     }
 
 	bool load(QFile& file);
 
-    void populate(QVector<QVector3D>& vertices, QVector<unsigned int>& indices, int maxDepth) const;
+    void populate(QVector<QVector3D>& vertices, QVector<unsigned int>& indices, size_t maxDepth) const;
 
 private:
-	unsigned int mDepth;
-	unsigned int mNodeCount;
-    unsigned int mInnerCount;
+	size_t mDepth;
+	size_t mNodeCount;
+	size_t mInnerCount;
 	struct Node* mRoot;
 	QVector3D mLowerBound;
 	QVector3D mUpperBound;
