+++ src/diagnostic/io/ImageBufferView.h
@@ -6,13 +6,13 @@ class ToneMapper;
 class ImageBufferView {
 public:
 	virtual const QString& name() const						 = 0;
-	virtual size_t width() const							 = 0;
-	virtual size_t height() const							 = 0;
-	virtual size_t channelCount() const						 = 0;
-	virtual size_t viewChannelCount() const					 = 0;
-	virtual const QString& channelName(size_t channel) const = 0;
+	virtual int width() const								 = 0;
+	virtual int height() const								 = 0;
+	virtual int channelCount() const						 = 0;
+	virtual int viewChannelCount() const					 = 0;
+	virtual const QString& channelName(int channel) const = 0;
 
-	virtual float value(size_t x, size_t y, size_t channel) const = 0;
+	virtual float value(int x, int y, int channel) const = 0;
 
 	void fillImage(QImage& image, const ToneMapper& mapper,
 				   quint32 channelOffset = 0, quint8 channelMask = 0xFF) const;
